Update LSUSB.h
Change type of usbTimeoutMillis from int to unsigned long to match millis() type

+++ Build_Files/Firmware_Files/LipSync_Firmware/LSUSB.h
@@ -182,7 +182,7 @@ void LSUSBMouse::begin(void)
 
   unsigned long timerHidTimeoutBegin = millis();
 
-  int usbTimeoutMillis;
+  unsigned long usbTimeoutMillis;
   usbAttempt++;
     
   if (usbRetrying) {
@@ -440,7 +440,7 @@ void LSUSBGamepad::begin(void)
 
   unsigned long timerHidTimeoutBegin = millis();
 
-  int usbTimeoutMillis;
+  unsigned long usbTimeoutMillis;
   usbAttempt++;
     
   if (usbRetrying) {
